Made a whole lot of changes
Waited too long to push
Recent changes include adding screen shake on going fast and landing from the air.
Updated the capsule size when in the air.
Lots of character changes including a new character model, texture, an entire rework of how the character handles making use of the skye BP now. Added a lot of component to offload the work so everything isn't in one monster file