Added various improvements
Made an update to the camera so that transitioning to sloped surfaces with manual camera isn't jittery. Also the camera follows better
Added glass decal and material.
Updated the Skye model and run animation.
Fixed the glitch when landing on edges that caused Skye to fall and land repeatedly.
Added some track segments with new textures.