Updated the way restarted levels works
Prior to this change, we simply restarted the level. This causes variable lag based on the amount of content in each level and would sometimes appear to freeze the game.

We updated the process so that it simply resets all the relevant data of the level to it's default state. Currently has a bug that causes the info page to appear each time you press accept which can lead to some of them being stuck on screen permanently. There are also some, likely, unrelated errors being thrown. Those should be fixed up in a different branch upon confirmation they are not related to level reset.