Updated the sense of speed without changing the actual speed
Updated the FoV to have a cap. This prevents the FoV from getting too crazy when going speeds well over max speed.
Updated the camera location using the spring arm. It zooms the camera in and lowers it closer to the ground at higher speeds. This has a negative side effect of being hard to see what is happening when going around ramps and corners.
Fixed a bug that was causing the player's speed to increase when in the air, while above max speed. This may need to be revisited as it currently doesn't affect the speed at all if above max speed.