Added a basic implementation of an infinitely generating track.
Creating a track that generates forever. This does cause the player to despawn once they reach the engine's range threshold of 1 million. Could probably increase that but physics gets wonky so a different approach may need to be implemented if I ever want that

Also implemented a debug feature that will enable spheres to indicate what is considered downhill or not