Playtest build created
Made a playtest build, realized that the performance was very bad on Steamdeck.
Rotating lights on the spray cans. Updated the way cans are rotated to be rotated on GPU and removed the lights
Added some new geometry to lower draw calls for buildings and implementing LODing for them
Turned off shadow casting on Skye and the crowds.