Fixed the 10k bug and added some polish
removed the node that was trying to shove Skye into the ground when catching a wall. This was sometimes causing her to get launched in a wrong direction after redirecting
Did a big overhaul to the level complete screen. Should be much more clear now. May still need to add clarity on where score is coming from and possibly how much more to the next rank?
Also updated the timer so that it has thousandths digits instead of just hundredths. This was inconsistent with what was being displayed on the level complete screen.